feat: allow multiple surveys
(Completed further study 2.)

In app.py:
* Moved the existing code at route "/" to "/survey/<survey_code>", so
that the root can instead display survey choices.
* In the route "/survey/<survey_code>", pointed template rendering to
renamed HTML file.  Used a dictionary to keep a bunch of survey info
together and only send needed info.  This route required a post
request, because it seemed like a better choice than having a route
"/surveys" with a query being the survey code.
* Modified the rest of the routes to include the survey code, because
it seemed like the proper way of structuring the URL.
* Stored the current survey's, that's being worked on, code in a Flask
session to keep track of the current survey.
* Changed the storing of responses to survey questions to being in a
dictionary in a Flask session, under the survey's code, in order to
store responses to multiple surveys.
* In the route "/survey/<survey_code>/questions/<int:ques_num>", added
logic for handling when the URL has a different survey code.
* In the route "/survey/<survey_code>/questions/<int:ques_num>", also
stored the data, being sent to render the template, in dictionaries for
better organization.
* In the route "/thankyou", extra variables were created, because it
will be used anyway in a future "further study".
* Modified the helper function __next_page__() accordingly with having
multiple surveys and storing responses differently.

In app.js:
* Created a JavaScript file to create a function that would cause the
survey selection form at the root page to go to the correct URL.

In base.html:
* Added JQuery.

For survey-start.html:
* Renamed index.html to survey-start.html, because index.html will now
be used to display survey choices.

In index.html:
* Created a new index.html with a survey selection form to allow users
to choose a survey.

In question.html:
* Updated argument variable use, passed from template rendering in
app.py.