feat: store cookies longer
(Completed further study 5.)

* Set the Flask session.permanent property to true, which stores Flask
session cookies for a default 31 days.  This had to be set soon after
storing something in a Flask session.

* Changed the code in the "/survey/<survey_code>/questions" route, so
that if there is previous stored session data, that is used, else
create and use a new List.

* Changed the returned redirect for the route
"/survey/<survey_code>/questions" to return a user to where the user
last left off, or to the first question of a survey.

* Made the button, at the introduction to a survey, responsive to a
user's previous position to the survey, by adding code that determines
the button text, at route "/survey/<survey_code>" in app.py, and using
the text in survey-start.html.