feat: allow revisiting previous questions
In question.html:
* Added a button to go to the previous question, and only display it if
the page isn't for the first question.

In app.py:
* Changed the logic for route
"/survey/<survey_code>/questions/<int:ques_num>", so that the currently
viewed question page is only invalid if the question number is out of
the range of the previous and current questions.  In other words, this
will now allow users to visit previous questions but not future
questions or negative question numbers.

* Changed route "/survey/<survey_code>/answer" to now append new
responses to the responses cookie data and replace old responses when
visiting previous questions.

* Changed route "/survey/<survey_code>/answer" to also redirect to the
previous question if the "previous" button is clicked, and to redirect
to the next incremented question page or the "thank you" page when the
"next" button is clicked.

* Renamed the helper function __next_page__ to __next_unvisited_page__
to clarify its purpose and avoid confusion with the redirect logic in
route "/survey/<survey_code>/answer".